apo types kala pame

+++ Part_1/search_ops.h
@@ -14,10 +14,10 @@
 //all of them in vectors with size=s_curves.size() each of them
 void search_curves(std::vector<real_curve> & s_curves,
   std::vector<std::vector<std::vector<entry>>> & Lht,
-  int k, int v_size, int dimension, int delta, int tablesize,
+  int k, int v_size, int dimension, double delta, int tablesize,
   std::string hash,std::string dist, std::vector<real_curve*> & pcurves,
   bool stats, double R, real_curve** nn_curve,double* nn_distance,
-  bool* grid_curve_found,std::vector<std::string> * curves_in_R, int w);
+  bool* grid_curve_found, std::vector<std::string> * curves_in_R, int w);
 
 //finds the (1)curve and (2)distance between the scurve and
 //the nearest curve in a vector of (pointers of) curves
